cloud microservices/cloud microservices/image/requirements.txt
Flask==1.1.2
gunicorn==20.0.4
Pyrebase4==4.4.3
